Can create Layer to GoogleMaps
To add Google Layers, we tried to include https://gitlab.com/IvanSanchez/Leaflet.GridLayer.GoogleMutant/-/blob/master/src/Leaflet.GoogleMutant.js?ref_type=heads but without success.
It does not seem a web-component, instead it is a JS that expands the existing Layer javascript class). So for the moment we simply create a TileLayer to the correct URL.